Update patch set 9
Patch Set 9:

There were big problems with handling argument with ';' char, because in older than 2.7.8 versions of python argparser stopped parsing on this char. Later versions see it as a part of an argument.
Did some simple check for this with a message.

Changed colouring for linux & now it should work on all versions.

Patch-set: 9
Label: Code-Review=0