[Win] Renderers: Introduce OpenGL 4 Renderer
This commit introduces Windows version of nfRendererOGL4 rendering
backend. Most of the functions are a stubs, however the main functionality
needed to initialize the renderer with OpenGL and further develop new
features to it is there. Things done:
  * Basic structure of the Renderer
  * OGL initialization (right now Backbuffer handles it, because of
    Device class design. At some point, the initialization should probably
    be moved to Device - this needs further checking at a later point)
  * Simple OGL Extension acquisition and handling (it should be enough for
    now)
  * Support for VSync to test OGL extensions

Additionally, nfRendererDemo was updated to handle currently unfinished
Renderer. Some functionalities were hidden behind flags and are called
only if Device returns a valid pointer to created interface. Moreover,
nfRendererDemo now accepts a launch argument - a name of renderer library
to use. If not provided, the default nfRendererD3D11 is used.

Verification:
    Run nfRendererDemo with nfRendererOGL4 option. A window with no contents
    should appear, VSync should be enabled, no errors should occur.

Change-Id: I5dc40a22e6641b33583136df3cc0ec4972c7c4e0