Update patch set 9
Patch Set 9:

I found due to my astonishment (as VAST number of programmers always did it using inline assembly) a header with cpuid implementation available for gcc.
I swaped it and pushed as draft. I'll be able to test it on my laptop, so if any of you have a chance to do it faster, please do so.

Patch-set: 9
Label: Code-Review=0