nfCommon: Introduce PackedArray template class
PackedArray is a class that holds all objects in a contiguous memory block,
even if they are created and removed in a random order.

PackedArray<ObjType, IDType> provides simillar functionality as
std::map<IDType, ObjType>, but all operations have O(1) average complexity.

The PackedArray class will be used to implement fast components creation
and access in Entity-Component-System in nfCore.

Verification:
    Rebuild and run nfCommonTests (you can filter only "PackedArray*")

Change-Id: Idd39b2702c339f150fe83057a6c0feb93c6b53a1