Renderers: Refactor resources binding [stage 1]
This is the first stage of resources binding refactor which is required
for introduction of future D3D12 and Vulkan renderers.
Only nfRendererD3D11, nfRendererDemo and nfRendererTest were modified
in this commit. nfCore will fail to build.

Breif description of the new binding mechanism:
Now, to bind resources to the rendering pipeline one needs two things:
1. Binding Set, which describes what resources types will be bound to
   a shader program at which slot (slot number can be obtain from
   IShaderProgram interface by name occurring in shader code).
2. Binding Instance, which is "implementation" of a Binding Set.
   It contains the list of actual resources that can be bound.
Binding Sets can be interpreted as C structures and Binding Instances
as the strucure objects. Binding Sets are groupped in Binding Layouts,
which enables a possibility to rebind only a part of the resources.
For example, one could create a following Binding Layout:
* binding set 0 with one cbuffer for vertex shader containing global info,
* binding set 1 with one cbuffer for pixel shader containing global info,
* binding set 2 with one cbuffer and 3 textures for pixel shader,
  containing material data, frequently updated.
This way, the binding set "2" can be updated frequenlty, when a material
changes, which rebinds cbuffer and all three textures in one go.

Major changes:
* Removal of SetTextures, SetSamplers and SetConstantBuffers method from
  ICommandList interface. They were replaced with single BindResources
  method.
* Introduction of new interfaces: IResourceBindingSet,
  IResourceBindingLayout and IResourceBindingInstance.
* Added needed objects in nfRendererDemo and rewritten resources binding.
* Implementation of the new binding mechanism in nfRendererD3D11.
* Each texture slot must have its sampler slot with the same number
  in the shader code.

Verification: Rebuild and run nfRendererDemo and nfRendererTest.

Change-Id: I1a3a6d5fd0d341c7ccf0334c4aab43c8b7b6f29d