nfCommon: Implement enum iterator class
Adds EnumIterator template class that exposes begin() and end() methods of
EnumIterator<*some enum class*>(), along with ++ and * operators, so that
range-based for is able to iterate over all enum class values and !=/==
operators for testing.

Simple test to nfCommonTest has been added as well.

Verification: Build and run tests.

Change-Id: I4545de9950b73f1dae8fec5ea0319e317cbc1e18