nfCommon: Implement Image::SaveJPG method
Introduces a way to save an object of the Image class to an OutputStream -
which can be either buffer or file. Also introduces tests for Save*()
methods.
Changes number of channels when loading JPG files to 3 - jpg format
doesn't use alpha channel.

Verification: Build and run nfCommonTest. Check for files in
nfEngine\TestResources\ImageSamples ending with "saved.jpg" and compare
them to originals.

Change-Id: I64003060acf48b4cc859b528b6dfa68eb6dce640