Update patch set 10
Patch Set 10: Code-Review-1

(9 comments)

I'm little terrified of xcb looking just at it's documentation (as it lacks A LOT).

Really nice, that X11 was implemented earlier, as I can see that it came in handy.

Please remove all calls to 'free()' that You are unsure of and check this Window thoroughly with valgrind.

Patch-set: 10
Label: Code-Review=-1