nfCore: Scene mamagement rewritten
* "Entity" is a normal class holding list of child entities and components.
* EntityManager was transformed to EntitySystem.
* TransformSystem was removed - EntitySystem is managing entities transforms.
* All the Scene Systems inherit from Scene::ISystem.
* Introduced concept of "GameObject". Game Object is a new resource
  representing a structure of nodes and components.
  Game Object can be spawned on a scene via GameObjectComponent.
  GameObjectSystem manages all spawned Game Object Instances.

RendererSystem and PhysicsSystem were refactored. They are now more
isolated from the rest of the code (scene and components), so they
can be moved to separate project in the future. Each of this
systems provides an interface to create "proxies" - the objects
managed by the system. They won't keep track of resources states -
it's components and resource manager job. For example, when a
MeshComponent is attached to entity, it will create MeshProxy
only when the Mesh resource is fully loaded. This will apply to
other systems as well.
NOTE: the proxy management interface is a draft - there will be more
features and proxy types required in the future.

Verification:
    1. Rebuild all.
    2. Run nfCoreTest and nfEngineDemo in all configurations.

Change-Id: I87209e0028bd06279076e7bd31f097e291966aba