nfCommon/DynArray: Don't allow for instering an element already contained by the container
Added an assert that will trigger if somebody tries to PushBack
an element that is already contained by this container. The reason
is that it can cause buffer reallocation and the reference to element
to be pushed would be invalid leading to UB or crash.

In the future we may want to add some additional code that recalculates
source element pointer if reallocation occurred.

Change-Id: Iec088b9225833dfa9b3559bb66e6e94ffb8da11d