RendererVk E09: Make Direct Binding work... somewhat
This change adds Direct Binding support to RendererVk. This originally
seemed like a great field for using Push Descriptors extension, however
each Pipeline Layout has a limit of sets allowing for Push Descriptors.
As an alternative, this commit uses "copy-and-overwrite" using temporary
Descriptor Sets.

Approach used is rather naive and leads to many unnecessary allocations
and deallocations (one alloc when direct binding and dealloc when
restarting/freeing a CommandRecorder).

This change revealed some flaws with Vulkan's descriptor set approach,
which makes scenes 0:6 and 0:9 not work after all. Underlying idea is
there, but to make it work properly some solid changes have to be made
to ResourceBinding.cpp and it's use of storing shader-defined Set Slot.