Add Core and EngineDemo builds in CMake
Core and EngineDemo were remains from previous commits which are now
properly configured and added to the build process.

EngineDemo is only added on Win32 because it uses WinMain and Windows
app subsystem, which makes it not buildable on other platforms. This
will require a bit of a rewrite once we get back to Core implementation.