RendererVk: Move VkShaderModules to PipelineState
This is preparation for new purely-direct resource binding model and is
a necessary step to allow remapping Descriptor Sets on the fly.

VkShaderModule is now owned by PipelineState and is created at the same
time as VkPipeline.

This adds a possibility to remove Shader objects after successfully
creating a PipelineState object.