CMakeLists: Use $ORIGIN as RPATH/RUNPATH on Linux build
While CMake-provided settings were working fine with DT_RPATH section
being used as default provider of dependency locations, newer systems
deprecate usage of DT_RPATH and switch to differently handled DT_RUNPATH
as default. This has an unintended side effect of not always working
with dlopen() and causing issues like RendererDemo not being able to
properly link to chosen Renderer backend.

This patch sets some CMake variables to make it use $ORIGIN as link path
instead of whatever default was chosen by CMake. As an added bonus, this
will now also work after moving the binaries to any location, as long as
required libraries are in the same directory as the executable.