RendererVk: Introduce DescriptorSetCache
This is a groundwork for future caching and reusing Descriptor Sets.
Also it helps with maintaining lifetime of Sets with some nifty
combinations and connections to CommandBufferManager.

For now the implementation is naive and frees Sets once Command Lists
using them were freed. That will change at some point (I hope).